feat: Update data types in config.py for consistency and accuracy
- Changed data types for keys in 'dim_product', 'dim_product_variant', 'dim_order', and 'fact_orders' from 'int' to 'float64' to accommodate larger values and ensure consistency across dimensions.

feat: Add Shopee Ads Data Extraction Script

- Introduced a new script `extract_shopee_ads.py` to extract monthly aggregate advertising data using the Shopee Ads API v2.
- Implemented error handling and logging for successful and failed extractions.

feat: Create Sample Ads Data Generator

- Added `generate_sample_ads.py` to create sample monthly aggregate ads data for testing purposes.
- Generates realistic sample data and saves it to a JSON file for pipeline testing.

feat: Implement Historical Review ID Extraction Method

- Created `extract_review_history_list` method in `temp_review_method.py` to extract review IDs using the Lazada API.
- Handles pagination and rate limiting while extracting historical reviews in 7-day chunks.

test: Add Tests for Updated Lazada Review Extraction

- Developed `test_lazada_reviews_updated.py` to test the new Lazada review extraction process using the updated API approach.
- Validates the extraction of review IDs and detailed review content.

test: Implement Shopee GMS API Single Month Test

- Created `test_shopee_gms.py` to test the Shopee GMS campaign performance for a single month.
- Validates data extraction and API usage reporting.

+++ app/config.py
@@ -486,7 +486,7 @@ COLUMN_DATA_TYPES = {
         'platform_key': 'int'
     },
     'dim_product': {
-        'product_key': 'int',
+        'product_key': 'float64',   
         'product_item_id': 'str',
         'product_name': 'str',
         'product_sku_base': 'str',
@@ -497,8 +497,8 @@ COLUMN_DATA_TYPES = {
         'platform_key': 'int'
     },
     'dim_product_variant': {
-        'product_variant_key': 'int',
-        'product_key': 'int',
+        'product_variant_key': 'float64',
+        'product_key': 'float64',
         'platform_sku_id': 'str',
         'variant_sku': 'str',
         'variant_attribute_1': 'str',
@@ -507,7 +507,7 @@ COLUMN_DATA_TYPES = {
         'platform_key': 'int'
     },
     'dim_order': {
-        'orders_key': 'int',
+        'orders_key': 'float64',
         'platform_order_id': 'str',
         'order_status': 'str',
         'order_date': 'datetime64[D]',  # Date only, not datetime
@@ -520,10 +520,10 @@ COLUMN_DATA_TYPES = {
     },
     'fact_orders': {
         'order_item_key': 'str',
-        'orders_key': 'int',
-        'product_key': 'int',
-        'product_variant_key': 'int',
-        'time_key': 'int',
+        'orders_key': 'float64',
+        'product_key': 'float64',
+        'product_variant_key': 'float64',
+        'time_key': 'float64',
         'customer_key': 'float64',  # Changed to float64 to match dim_customer (1.1, 1.2, etc.)
         'platform_key': 'int',
         'item_quantity': 'int',
